src/main/java/ru/aafanasiev/util/copier/converter/clocaldate/CopyConverterOffsetDateTime2SDate.java
/**
 *
 */
package ru.aafanasiev.util.copier.converter.clocaldate;

import java.sql.Date;
import java.time.OffsetDateTime;

import ru.aafanasiev.annotations.Converter;
import ru.aafanasiev.util.copier.converter.base.ConverterKey;
import ru.aafanasiev.util.copier.converter.base.CopyConverter;
import ru.aafanasiev.util.copier.converter.base.CopyEntry;

/**
 * @author aafanasyev
 */
@Converter
public class CopyConverterOffsetDateTime2SDate implements CopyConverter {

    /*
     * (non-Javadoc)
     * @see ru.aafanasiev.util.copier.converter.base.CopyConverter#convert(ru.aafanasiev.util.copier.converter.base.
     * CopyEntry, java.lang.Object)
     */
    @Override
    public Object convert(CopyEntry entry, Object value) {
        if(null != value && OffsetDateTime.class.equals(value.getClass())) {
            return new java.sql.Date(Date.from(((OffsetDateTime) value).toInstant()).getTime());
        }
        return null;
    }

    /*
     * (non-Javadoc)
     * @see ru.aafanasiev.util.copier.converter.base.CopyConverter#listConvertedPairs()
     */
    @Override
    public ConverterKey[] listConvertedPairs() {
        return new ConverterKey[] {new ConverterKey(OffsetDateTime.class, Date.class)};
    }

}
